+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 Below should serve as an "upgrade guide" jumping off point as you are migrating projects to newer versions of this package. Ideally, interface changes will be called out specifically and one should not have to go sifting through git diffs to discern what changed.
 
 - `?.?.?` - *WIP*
+   - Added methods:
+      - `mapP`
+   - Mehod name change:
+      - `timeout` -> `timeoutP`
 - `1.1.8`
    - Ability to generate a static documentation site that's hosted on GitHub pages.
 - `1.1.7`